TaskMgr/TaskMgrTgBot/requirements.txt
pydantic==2.3.0
nest-asyncio==1.5.5
aiogram==3.1.1
aiogram_dialog==2.0.0
requests==2.28.2
emoji==2.2.0
python-dotenv==1.0.0
aiohttp~=3.8.6
python-odata==0.5.4